Add install and uninstall subcommands for Windows
On Windows 10+, there are specific registry entries you need to create
in order to register a program as a browser. Otherwise, you cannot
select it as the default browser in the "Default apps" section of the
control panel.

The install subcommand creates these registry entries and, if provided
with the --set-default flag, will open the Windows Control Panel so the
user can set the default browser. Windows does not expose a explicit
"set-as-default" API on purpose, and while it is possible to do some
clever registry editing (à la Firefox) to override the default browser,
the use case of metabrowser is niche enough that I can't be bothered to
do so.