Build as a Windows subsystem application
By building as a Windows subsystem application, we no longer show a
console Window when opening from a non-console Window context (e.g.,
when metabrowser is the default browser and a link is opened with it).
However, as a Windows subsystem application, the console is by default
detached, so any text printed won't appear, even when invoked from the
command-line. We attach to the parent process' console so that we can
still print, e.g., error messages and instructions when calling
`metabrowser install --set-default`.