Match more carefully against TLDs and paths in URLs
Previously, a URL rule look `example.com` would also match
`example.com.evil.com`. Likewise, a URL rule with a partial path, like
`example.com/path` would match `example.com/path-to-evil-subdir`. We now
add a suffix to each pattern to ensure either (a) the URL being matched
ends at the end of the pattern, or (b) the next character of the URL is
a `/`, so we will also match subdirectories.